[tz-459] Fix circular dependency issue preventing render module from building in workflow to bump and deploy libs automatically.
